feat(form): Create utils file to have general form methods
